+++ ChangeLog.md
@@ -5,6 +5,8 @@ ChangeLog {#changelog}
 - Command line tools are now placed in the `bin` directory of modules (the
   old `bin` directory, used for utility programs that are not installed, is
   now called `utility`). The `applications` directory is removed.
+- Several Python modules that were only used by EMageFit have been moved from
+  IMP::em2d to IMP::EMageFit.
 
 # For IMP 2.3.0 # {#changelog_2_3_0}
 - A new IMP::pmi module was added, containing high-level Python classes to simplify the representation, scoring, sampling and analyzing of models in IMP. This module is still under heavy development, and should be considered experimental.